Change API to allow different instances
Previously, highlight.js was a global singleton.
So was lowlight.
If you registered a language in one of them, it was available in the rest.

It isn’t anymore, which means we don’t need a single global lowlight instance
anymore either.

This removes the previous `lowlight` export.
You can now use the `createLowlight` export to create one.

This removes the `all.js` and `common.js` files.
There are now `all` and `common` exports, which you can pass to `createLowlight`,
so as to register nothing (`undefined`), common languages (`common`), or
everything (`all`).